fix(loans): move debt-analysis charts out of pages/ to fix Vercel prerender
Components living under src/pages/loans/_*/index.tsx were being treated
as routes by Next.js (`/loans/_DebtCurveChart`, `/loans/_CheapestBankChart`)
and prerendering tried to call them with no props — failing on the
useMemo .forEach() calls.

Move both to src/components/loans/, matching LoansBlotterTable's
existing convention.
